Add YouTube like animiation.
Submit the PR as of now because the animation works similarly as
the YouTube app.
But ideally only the height of the animated image should change first
until the image moves close to the bottom of the screen, then the
width shrinks after that.

But looks like it's not supported to interpolate the size of an view
using KeyAttribute or KeyPosition.

I think if MotionLayout allows an intermediate ConstraintSet, the
animation becomes possible.